Change key bindings on function layer
Remove the navigation keys on WASD and add Ctrl, Alt and Win on SDF keys
to provide more key combinations closer to the home row.

Also, swap volume control keys on the function layer to put mute on the
M key and volume down and volume up on the <> keys to be more intuitive.