Remove Del from funcion layer and swap lcntrl for lshift
I want the word-delete key combination (Command-backspace) on the
function layer. Backspace is currently Delete on the function layer,
which is not useful to me. Remove Delete so that the Backspace key is
Backspace on the function layer.

Replace lcntrl with lshift on the function layer so that it is possible
to do things like word highlighting in text editors and word jumping in
vim from the home row.